remove vitest dependency (#12)
Summary:
Pull Request resolved: https://github.com/facebook/openzl/pull/12

The latest Vitest release has a transitive dependency on an older version of Vite with a vulnerability. This is fixed on dev, but a release hasn't been cut. Until then, we won't be using vitest to test. They're working on 4.0.0 right now, so we will wait for a stable version of their next big update.

Reviewed By: terrelln

Differential Revision: D83600072

fbshipit-source-id: 8f4f9f7c12cbee6765f6fe9c3231b397be323dcd