Check in Pareto optimal results (#11)
Summary:
Pull Request resolved: https://github.com/facebook/openzl/pull/11

These results, along with the files run on, were uploaded to a GitHub [release](https://github.com/facebook/openzl/releases/tag/openzl-sample-artifacts). But we should also check in the results we got, as well as the command lines run, so they don't get lost & we can present a single release tag.

Reviewed By: Victor-C-Zhang

Differential Revision: D83597702

fbshipit-source-id: 8c7db89bae48e0f537269d682c74da176449855a