Selectors can select Segmenters (#53)
Summary:
`Segmenters` **must** operate on *the entire* user `inputs`. For this reason, they are meant to run first, before any other `graph` or `node`.

However, "other `graphs`" include `selectors`.
Now, `selectors` don't modify `inputs`. Therefore, it should be possible for a `selector` to run first, in order to select which `segmenter` to invoke. This property can be extended to any function graph that in fine only select.
This capability is important for a future in which a ML-driven `selector` could detect which kind of `input` is being processed in order to select the proper `segmenter` to deal with it.

This patch makes this scenario possible.

Pull Request resolved: https://github.com/facebook/openzl/pull/53

Reviewed By: terrelln

Differential Revision: D83875019

Pulled By: Cyan4973

fbshipit-source-id: 57f19d7b0a9dc87c0a2c10b168c86f7717b03356