add linting (#87)
Summary:
Pull Request resolved: https://github.com/facebook/openzl/pull/87

This diff adds ESLint and prettier to the visualizer repo. Updated the readme as well to reflect new steps to ensure clean lints.
For those working within VSCode, also added configs to save on format and highlight linted lines.

This does **not** add linting to the CI. There are a number of failures that need to be addressed before we can start blocking lands for lint errors.

Reviewed By: kevinjzhang

Differential Revision: D83709908

fbshipit-source-id: 36acd4c784aabf945db5d5a07dd1d54236a51e31